+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
  * added `--parsed` option to `puli config`
  * added `--format` option to `puli package`
  * added `puli package rename` command
+ * added required argument `server` to `puli asset map`
 
 * 1.0.0-beta4 (2015-04-13)